ODY-172 consolidated interface BlockNoteBlock
